+++ src/routes/index.js
@@ -8,12 +8,12 @@ const image = require('../controllers/image' );
 
 
 module.exports = app => {
-    router.get('/', home.index);
-    router.get('/images/:image_id', image.index);
-    router.post('/images', image.create);
-    router.post('/images/image_id/like', image.like);
-    router.post('/images/image_id/comment', image.comment);
-    router.delete('/images/image_id', image.remove);
+    router.get('/', home.index);                                // Página de inicio
+    router.get('/images/:image_id', image.index);               // Página de imagen
+    router.post('/images', image.create);                       // Subir imagen
+    router.post('/images/image_id/like', image.like);           // Like
+    router.post('/images/image_id/comment', image.comment);     // Comment
+    router.delete('/images/image_id', image.remove);            // Borrar imagen
 
     app.use(router);
 };