+++ src/controllers/home.js
@@ -5,8 +5,11 @@
 
 const ctrl = {};
 
-ctrl.index = (req, res) => {
-  res.render('index');
+const { Image } = require('../models');
+
+ctrl.index = async (req, res) => {
+  const images = await Image.find().sort({ timestamp: -1 });
+  res.render('index', { images });
 };
 
 module.exports = ctrl;