+++ tutorial/Red social Imgur - Proyecto terminado/src/controllers/image.js
@@ -13,11 +13,11 @@ ctrl.index = async (req, res) => {
   let viewModel = { image: {}, comments: [] };
   const image = await Image.findOne({filename: { $regex: req.params.image_id }});
   if (image) {
-    image.views = image.views + 1;
+    image.views += 1;
     viewModel.image = image;
     image.save();
     const comments = await Comment.find({image_id: image._id})
-      .sort({'timestamp': 1});
+      .sort({timestamp: 1});
     viewModel.comments = comments;
     viewModel = await sidebar(viewModel);
     res.render('image', viewModel);