+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## ... Vistas home y layout. Bug view [x]
+- Aplicación de vistas
+  - Se añade vista home con layout main (cabecera con logo)
+  - Uso de Bootstrap cpmp framework CSS
+  - Uso de fontsawesome.com como proveedor de iconos
+  - Modificaciones peronalizadas de estilo en archivo */public/stiles.css*
+- Solucionado BUG en el que estaba mal configurado el directorio de vistas en express y no redireccionaba bien.
+
+
 ## v.1.0.1
 - Cambiado enrutamiento, usando objeto *Express.Router()*
 - Se agrega uso de controladores en el direccionamiento.